Line 28 - Recommend splitting this paragraph
Separate the first sentence from the second and third.  As currently written it appears the last 2 sentences also apply to the Enterprise discussed in the first.  
Adding "publicly trusted" before "chain of trust" in 2nd sentence makes it clear that this is the intent. 
Consider adding ". . ., nor do they address the issuance or management of certificates by enterprises that interoperate via a Bridge CA environment." to the end of the first sentence to cover this credential use case - not exclusively internal but also not publicly-trusted.  Bridges are discussed elsewhere, this reference is in keeping with that.