Ballot SMC08 - Deprecate Legacy Generation Profiles and Minor Updates  (#253)
The S/MIME Baseline Requirements include a set of Legacy profiles designed to ease the transition into an audited framework, with the expectation that those Legacy profiles would be deprecated. This ballot sets a date by which issuance of certificates following the Legacy generation profiles must cease. It also includes the following minor updates:

Pins the domain validation procedures to v 2.0.5 of the TLS Baseline Requirements while the ballot activity for multi-perspective validation is concluded, and the SMCWG determines its corresponding course of action;

Updates the reference for SmtpUTF8Mailbox from RFC 8398 to RFC 9598; and

Small text corrections in the Reference section

This ballot is proposed by Stephen Davidson (DigiCert) and endorsed by Clint Wilson (Apple) and Martijn Katerbarg (Sectigo).